nexus_wallet_package/requirements.txt
# NexusOS Web3 Wallet Dependencies
# =================================

# Web3 blockchain integration
web3>=7.0.0
eth-account>=0.13.0
eth-keys>=0.7.0
hexbytes>=1.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# UI/UX
streamlit>=1.20.0
streamlit-autorefresh>=1.0.0
plotly>=5.0.0

# Data processing
numpy>=1.20.0
pandas>=1.3.0
scipy>=1.7.0

# CLI interface
click>=8.0.0
rich>=13.0.0

# Configuration
python-dotenv>=1.0.0

# Security
bcrypt>=4.0.0
passlib>=1.7.0

# Utilities
requests>=2.28.0
